- Modifiquei o tipo do valor de amount para BigDecimal
- Adicionei novas funcoes para somar o total do valor das despesas(total, fixa ou variavel) receita e etc

- Adicionei enums para o metodo de pagamento e para categoria

- Adicionei a propriedade de salario para o usuario, para poder ser feito os calculos

+++ SGD/backend/src/main/java/com/klaus/backend/Service/RelatorioService.java
@@ -223,8 +223,8 @@ public class RelatorioService {
                 // Valor (número)
                 Cell c4 = row.createCell(col++);
                 if (e.getAmount() != null) {
-                    c4.setCellValue(e.getAmount());
-                    totalValor = totalValor.add(BigDecimal.valueOf(e.getAmount()));
+                    c4.setCellValue(e.getAmount().doubleValue());
+                    totalValor = totalValor.add(e.getAmount());
                 } else {
                     c4.setCellValue(0.0);
                 }
